Select All on sites no longer sorts. Sites table resizes appropriately. Significant restructuring to sites' divs.
The Select All on the sites table used to sort the records. It no longer does so.

The sites table would overflow past the container if the browser were sized smaller. Now it doesn't.

Previous placement of divs negatively affected attempts to style select items. This commit should alleviate some of those weaknesses.