Improved UX on site filters.
Previously, failing to select an option for the filter would be like passing a value like null or undefined into the filter (found in sites.js). This would break the table.

Now, not selecting an option for the filter simply selects both options for the respective filter. For example, not selecting a program filter and pressing the filter button will return both Bio and Itec records.